requirements.txt
# 核心框架
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0

# 音频处理
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
numpy>=1.24.0
scipy>=1.10.0

# 机器学习
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 音频模型
# RVC相关
faiss-cpu>=1.7.4
praat-parselmouth>=0.4.3
pyworld>=0.3.0
scikit-learn>=1.3.0

# Web框架
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# 工具库
pyyaml>=6.0
requests>=2.31.0
pillow>=10.0.0
tqdm>=4.66.0

# 数据库
sqlalchemy>=2.0.0

# 配置管理
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 类型检查
pydantic>=2.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
